ci: preserve survivor coverage within existing limits
Keep all native operator and schema cases in their existing owner suites. Register the shell observer entry and current inert scenario catalog, and update fixture, request, and routing contracts without changing runner caps or assertions.
